+++ examples/basic/__tests__/App-test.js
@@ -4,11 +4,11 @@
 
 import 'react-native';
 import React from 'react';
-import VideoPlayer from '../src/VideoPlayer';
+import App from '../App';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  renderer.create(<VideoPlayer />);
+  renderer.create(<App />);
 });